Struct soapysdr::Args

pub struct Args(_);

A list of key=value pairs.

Implementations

impl Args

pub fn new() -> Args

Create a new, empty Args list

pub unsafe fn from_raw(a: SoapySDRKwargs) -> Args

Safety

Be careful that SoapySDRKwargs is either:

SoapySDRKwargs::keys and SoapySDRKwargs::vals are null and SoapySDRKwargs::size is 0 or

SoapySDRKwargs::keys and SoapySDRKwargs::vals both point to valid keys and vals of SoapySDRKwargs::size length.

pub fn as_raw(&mut self) -> *mut SoapySDRKwargs

pub fn as_raw_const(&self) -> *const SoapySDRKwargs

pub fn set<K: Into<Vec<u8>>, V: Into<Vec<u8>>>(&mut self, key: K, value: V)

Append a key-value pair to the arguments list

Example

use soapysdr::Args;
let mut args = Args::new();
args.set("driver", "lime");

Panics

if key or value contain null bytes

pub fn get<'a>(&'a self, key: &str) -> Option<&'a str>

Get the value corresponding to a key in the arguments list.

Example:

use soapysdr::Args;
let args: Args = "serial=123456".into();
assert_eq!(args.get("serial"), Some("123456"));

pub fn iter(&self) -> ArgsIterator<'_>

Get an iterator over the (key, value) pairs in the arguments list.

Example:

use soapysdr::Args;
let args: Args = "driver=lime, serial=123456".into();
let mut i = args.iter();
assert_eq!(i.next(), Some(("driver", "lime")));
assert_eq!(i.next(), Some(("serial", "123456")));
assert_eq!(i.next(), None);