Trait soapysdr::StreamSample

pub unsafe trait StreamSample {
    const STREAM_FORMAT: Format;
}

Trait for sample formats used by a TxStream or RxStream

Safety

Implementing this trait requires that the type have the same size, alignment, and compatible memory representation with the SoapySDR type selected by STREAM_FORMAT

Required Associated Constants

const STREAM_FORMAT: Format